Remove the need to refer to `::extract::`, or `::prepare::`, or `::write::`.
