+++ demo/src/features/debug3d/public/mod.rs
@@ -1,1 +1,2 @@
-pub mod debug3d_resource;
+mod debug3d_resource;
+pub use debug3d_resource::*;
