+++ demo/src/features/imgui/mod.rs
@@ -1,21 +1,21 @@
 use rafx::render_feature_mod_prelude::*;
 use rafx::render_feature_renderer_prelude::*;
-
-use distill::loader::handle::Handle;
-use extract::ExtractJobImpl;
-use internal::*;
-use rafx::assets::MaterialAsset;
-use rafx::framework::{ImageViewResource, ResourceArc};
-
 rafx::declare_render_feature!(ImGuiRenderFeature, DEBUG_3D_FEATURE_INDEX);
 
 mod extract;
+use extract::*;
 mod prepare;
+use prepare::*;
 mod write;
-
+use write::*;
 mod internal;
+use internal::*;
 mod public;
-pub use public::Sdl2ImguiManager;
+pub use public::*;
+
+use distill::loader::handle::Handle;
+use rafx::assets::MaterialAsset;
+use rafx::framework::{ImageViewResource, ResourceArc};
 
 struct StaticResources {
     pub imgui_material: Handle<MaterialAsset>,
@@ -30,7 +30,7 @@ impl ImGuiRendererPlugin {
         resources: &mut legion::Resources,
         window: &sdl2::video::Window,
     ) {
-        let imgui_manager = public::init_sdl2_imgui_manager(window);
+        let imgui_manager = init_sdl2_imgui_manager(window);
         resources.insert(imgui_manager);
     }
