+++ demo/src/features/imgui/public/mod.rs
@@ -1,3 +1,3 @@
 mod sdl2_imgui_manager;
-pub use sdl2_imgui_manager::init_sdl2_imgui_manager;
+pub(super) use sdl2_imgui_manager::init_sdl2_imgui_manager;
 pub use sdl2_imgui_manager::Sdl2ImguiManager;
