+++ demo/src/features/imgui/public/sdl2_imgui_manager.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 use std::sync::Mutex;
 
-use crate::features::imgui::internal::{ImGuiDrawData, ImguiFontAtlasData, ImguiManager};
+use super::super::{ImGuiDrawData, ImguiFontAtlasData, ImguiManager};
 use imgui_sdl2::ImguiSdl2;
 use sdl2::mouse::MouseState;
 use sdl2::video::Window;
@@ -276,7 +276,7 @@ fn init_imgui(window: &Window) -> imgui::Context {
     imgui
 }
 
-pub fn init_sdl2_imgui_manager(window: &Window) -> Sdl2ImguiManager {
+pub(in super::super) fn init_sdl2_imgui_manager(window: &Window) -> Sdl2ImguiManager {
     let imgui_context = init_imgui(&window);
     Sdl2ImguiManager::new(imgui_context, window)
 }
