+++ demo/src/features/sprite/public/sprite_render_node_set.rs
@@ -1,4 +1,4 @@
-use crate::features::sprite::render_feature_index;
+use super::super::render_feature_index;
 use distill::loader::handle::Handle;
 use rafx::assets::ImageAsset;
 use rafx::base::slab::{DropSlab, DropSlabKey};
