+++ demo/src/features/text/internal/mod.rs
@@ -1,3 +1,6 @@
 mod font_atlas_cache;
 pub use font_atlas_cache::FontAtlasCache;
 pub use font_atlas_cache::TextImageUpdate;
+
+mod text_draw_data;
+pub use text_draw_data::*;
